# specs/001-rpg-social-media/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: RPG Social Media API
  description: |
    Twitter-like social media platform with RPG gamification mechanics.
    Users earn XP through interactions, level up, and unlock features progressively.
  version: 1.0.0
  contact:
    name: RPG Social Media Team
servers:
  - url: https://api.rpg-social.workers.dev
    description: Production (Cloudflare Workers)
  - url: http://localhost:8787
    description: Local development (wrangler dev)
security:
  - cookieAuth: []
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user account
      description: |
        Creates a new user account with default level 1 and 0 XP.
        Returns JWT token in httpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                  pattern: '^[a-zA-Z0-9_]+$'
                  example: "player123"
                password:
                  type: string
                  minLength: 8
                  example: "securePassword123"
      responses:
        '201':
          description: User created successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: auth_token=eyJhbGc...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login to existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "player123"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout current user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: auth_token=; Max-Age=0
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Users ====================
...
  /api/feed/home:
    get:
      tags: [Feed]
      summary: Get personalized home feed
      description: |
        Returns posts from followed users in reverse chronological order.
        Success Criteria: <2s load time for 100 followed users (SC-008)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Home feed posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostWithAuthor'
                  has_more:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/feed/discover:
    get:
      tags: [Feed]
      summary: Get discovery feed
      description: Returns popular posts for new users or exploration
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Discovery feed posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostWithAuthor'
                  has_more:
                    type: boolean

  # ==================== XP & Levels ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
      description: Number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of items to skip
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 100
        total_xp:
          type: integer
        xp_for_current_level:
          type: integer
          description: XP required to reach current level
        xp_for_next_level:
          type: integer
          description: XP required to reach next level
        xp_progress_percent:
          type: number
          format: float
          description: Progress to next level (0-100)
        created_at:
          type: integer
          description: Unix timestamp (milliseconds)
        avatar_url:
          type: string
          format: uri
          nullable: true
        banner_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        theme_preference:
          type: string
          enum: [default, dark, light, rpg, cyberpunk]
        # Character stats
        total_posts:
          type: integer
        total_likes_given:
          type: integer
        total_likes_received:
          type: integer
        total_comments_made:
          type: integer
        followers_count:
          type: integer
        following_count:
          type: integer
    PostWithAuthor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        content:
          type: string
        char_count:
          type: integer
        media_url:
          type: string
          nullable: true
        created_at:
          type: integer
        is_pinned:
          type: boolean
        author:
          type: object
          properties:
            username:
              type: string
            level:
              type: integer
            avatar_url:
              type: string
              nullable: true
        like_count:
          type: integer
        comment_count:
          type: integer
        is_liked_by_user:
          type: boolean
          description: True if authenticated user has liked this post
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...